Version 2.0.0

+++ src/main/java/net/dirtydeeds/discordsoundboard/async/CleanBotMessagesJob.java
@@ -2,6 +2,7 @@ package net.dirtydeeds.discordsoundboard.async;
 
 import java.time.OffsetDateTime;
 import java.time.ZoneId;
+import java.util.Date;
 import java.util.LinkedList;
 import java.util.List;
 
@@ -17,20 +18,21 @@ import net.dv8tion.jda.utils.SimpleLog;
 public class CleanBotMessagesJob extends AbstractAsyncJob {
 
     public static final SimpleLog LOG = SimpleLog.getLog("CleanBotMessagesJob");
+    private static final long MINUTES = 60000; // 1 minute in ms
+    private static final String SERVER_MESSAGE_PREFIX = "**FYI**";
 	
 	public void handle(SoundboardDispatcher dispatcher) {
-		LOG.info("Initiating job.");
 		MessageHistory history;
 		List<Message> messages;
 		List<Guild> completedServers = new LinkedList<>();
 		OffsetDateTime lastCheckedTime = (pastEvent != null) ? OffsetDateTime.ofInstant(pastEvent.time.toInstant(), ZoneId.systemDefault()) : null;
+		OffsetDateTime tooRecent = OffsetDateTime.ofInstant(new Date(System.currentTimeMillis() - 5*MINUTES).toInstant(), ZoneId.systemDefault()); // 5 minutes ago
 		for (SoundboardBot bot : dispatcher.getBots()) {
 			long deleted = 0;
 			JDA api = bot.getAPI();
 			if (api == null) continue;
 			for (Guild guild : bot.getGuilds()) {
 				if (!completedServers.contains(guild)) {
-					LOG.info("Starting to clean messages for server " + guild.getName());
 					// Clean bot's messages from this guild's channels.
 					for (MessageChannel channel : guild.getTextChannels()) {
 						history = new MessageHistory(channel);
@@ -41,11 +43,12 @@ public class CleanBotMessagesJob extends AbstractAsyncJob {
 									break channelloop;
 								}
 								if (msg.getAuthor() != null && bot.isUser(msg.getAuthor()) && 
-										(msg.getEmbeds() == null || msg.getEmbeds().isEmpty())) {
-									// This is a bot message and contains no embeds.
+										(msg.getEmbeds() == null || msg.getEmbeds().isEmpty()) &&
+										!msg.getContent().startsWith(SERVER_MESSAGE_PREFIX) &&
+										msg.getTime().isBefore(tooRecent)) {
+									// This is a bot message and contains no embeds, is not a server message, and is not a recent message.
 									try {
 										msg.deleteMessage(); ++deleted;
-										LOG.info("Deleted message with content: \"" + msg.getContent().substring(0,Math.min(msg.getContent().length(), 10)) + "...\".");
 									} catch (Exception e) {
 										LOG.fatal("While deleting message " + msg + " ran into exception.");
 										e.printStackTrace();
@@ -56,11 +59,10 @@ public class CleanBotMessagesJob extends AbstractAsyncJob {
 							catch (Exception e) { continue; }
 						}
 					}
-					LOG.info("Finished server " + guild.getName());
 					completedServers.add(guild);
 				}
 			}
-			LOG.info("Deleted " + deleted + " messages for bot " + bot.getName());
+			if (deleted > 0) LOG.info("Deleted " + deleted + " messages for bot " + bot.getBotName());
 		}
 	}
